Step-5-TrainingOverBLE/training/requirements.txt
numpy
gym
keras-rl
theano
pyserial